# rooiam-sdk/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rooiam API",
    "description": "Self-hosted passwordless IAM for multi-tenant SaaS.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/setup/auth-methods": {
      "get": {
        "tags": [
          "browser"
        ],
        "operationId": "get_public_auth_methods",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "org",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspace",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "client_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "widget_embed_origin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login methods enabled for the resolved workspace (public; no auth)"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/v1/setup/login-bootstrap": {
      "get": {
        "tags": [
          "browser"
        ],
        "operationId": "get_login_bootstrap",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "org",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspace",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "client_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "widget_embed_origin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Everything the hosted login widget needs to render: auth methods + workspace branding (public; no auth)"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "oidc_access_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "OIDC access token from /oidc/token, sent as `Authorization: Bearer <token>`."
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rooiam_session",
        "description": "Opaque session cookie set on magic-link verify; sent automatically by the browser."
      },
      "workspace_api_key": {
        "type": "http",
        "scheme": "bearer",
        "description": "Workspace API key, sent as `Authorization: Bearer <key>`."
      }
    }
  },
...
}